Move `lib` to `src` and rewire babel output directory.
This allows for cherry-picking modules, particularly the asset server middleware. Previously this what not run through babel, causing errors.

This also adjusts the eslint configuration to accomodate.
