+++ .eslintignore
@@ -1,0 +1,5 @@
+node_modules/**/*
+lib/**/*
+runtime/**/*
+bin/**/*
+*.js
