+++ .gitignore
@@ -1,3 +1,8 @@
 dist
 node_modules
 *.log
+/lib
+/bin
+/*.js
+/runtime
+*.map
