+++ src/lib/compiler.js
