+++ src/lib/platform/node.js
